Cleaned up code , moved Modal to separate file, cleaned up dumb component to not contain any redux state
