+++ app/containers/FileManagerPage.jsx
@@ -1,14 +1,14 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
+import React, {Component} from 'react';
+import {connect} from 'react-redux';
+import {bindActionCreators} from 'redux';
 import * as actions from '../redux/actions/index';
 import adb from '../utils/adb';
 import AppBar from 'material-ui/AppBar';
 import NavLeft from 'material-ui/svg-icons/navigation/chevron-left';
-import { goBack, push } from 'react-router-redux';
+import {goBack, push} from 'react-router-redux';
 import _ from 'lodash';
 import ContentList from '../components/FileManager/ContentList';
-import HeaderIcons from '../components/Header/Buttons';
+import HeaderIcons from '../components/Header/FileManagerButtons';
 
 const styles = {
   navLeftButton: {
@@ -37,15 +37,25 @@ class FileManagerPage extends Component {
     const selectedDeviceName = _.result(vm.props.state.devices, 'selectedDevice.device.properties["ro.product.model"]') || vm.props.state.devices.selectedDevice.serial;
     return (
       <div>
-        <AppBar title={`File Manager (${selectedDeviceName})`} iconElementRight={<HeaderIcons onRefreshClick={() => vm.loadFilesOfDirectory()}/>} iconElementLeft={<NavLeft style = {
-      styles.navLeftButton
-      }
-      onClick = {
-      () => vm.props.navigateBack()
-      } />}/>
+        <AppBar
+          title={`File Manager (${selectedDeviceName})`}
+          iconElementRight={
+            <HeaderIcons onRefreshClick = {
+                () => vm.loadFilesOfDirectory()
+              }/>
+            }
+          iconElementLeft={
+            <NavLeft
+              style={styles.navLeftButton}
+              onClick={
+                  () => vm.props.navigateBack()
+              }>
+            </NavLeft>
+          }>
+          </AppBar>
         <ContentList uiConfig={vm.props.state.filemanager.uiConfig} currentPath={vm.props.state.filemanager.currentPath} fileList={vm.props.state.filemanager.fileList}></ContentList>
       </div>
-      );
+    );
   }
 }
 
@@ -55,9 +65,7 @@ FileManagerPage.propTypes = {
 };
 
 let mapStateToProps = (state) => {
-  return {
-    state: state
-  };
+  return {state: state};
 };
 
 let mapDispatchToProps = (dispatch) => {
