+++ app/redux/actions/index.js
@@ -4,7 +4,6 @@ export const DEVICES_SELECT_DEVICE = 'DEVICES_SELECT_DEVICE';
 export const DEVICES_UPDATE_DEVICE_PROPERTIES = 'DEVICES_UPDATE_DEVICE_PROPERTIES';
 export const FILEMANAGER_CHANGE_PATH = 'FILEMANAGER_CHANGE_PATH';
 export const FILEMANAGER_CHANGE_UI_CONFIG = 'FILEMANAGER_CHANGE_UI_CONFIG';
-export const ADD_WIRELESS_DEVICE = 'ADD_WIRELESS_DEVICE';
 
 //DEVICES_ACTION_CREATORS
 export function updateDevices(devices) {
@@ -14,12 +13,6 @@ export function updateDevices(devices) {
   };
 }
 
-export function addWirelessDevice(ip) {
-  return {
-    type: ADD_WIRELESS_DEVICE,
-    ip
-  };
-}
 export function selectDevice(deviceSerial, device) {
   return {
     type: DEVICES_SELECT_DEVICE,
